Move SessionMode and LockMode to seata-common and apply StoreMode from seata-common
